Initialised completed project that converts raw text file (.txt) to .epub format, (specifically for converting chinese novel raw text into epub for reading on Apple iBooks) This is done by parsing the file using the parse function from parser, which returns the book name, a list of chapter titles, as well as a list of chapter contents (which are also lists of strings themselves).
Then, the respective files required for an epub are created in the
various functions in main.py, and finally, these files are zipped
together to form a epub file in the zipping area directory, and
moved to the current working directory.

This project was created because as an avid reader of chinese novels,
the downloads of quality novels in epub format were being removed
from the websites slowly but surely, and converting in on a text to Epub
site was not only time-consuming -- there was a severe limitation on
the number and size of files I could convert, not to mention the lack
of flexibility in modifying the style (css) of the novel resulting
in poor formatting that could not be changed in the novel, leading to a
poor reading experience for me.
Hence, I decided to cut out the middleman and write my own script that
converted txt to epub, as the various txt to epub modules I could find
online were either outdated (python 2), esoteric and difficult to use,
or extremely slow (30min-2h), which was unacceptable as I could not
comprehend how parsing a simple text file and converting it to epub
format could take so long.
It was kinda suspicious, and I didn't want to risk malware on my
computer nor entertain the possibility of my computer being used
to mine cryptocurrency (as my computer was overheating at times.), so
I embarked on this difficult journey of converting txt to epub myself.
Alas, it paid off as this project is working for it's intended use and
can be modified to convert other txt files to epub with a little
knowledge of Python.